fix norminette stuff
