+++ src/bonus/render_bonus.c
@@ -6,7 +6,7 @@
 /*   By: jorvarea <jorvarea@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/11 17:56:27 by ana-cast          #+#    #+#             */
-/*   Updated: 2025/06/13 17:04:14 by jorvarea         ###   ########.fr       */
+/*   Updated: 2025/06/13 18:15:16 by jorvarea         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,11 +32,12 @@ bool	stop_condition(t_map_tile **mt, double position[2],
 void	select_texture(t_game *game, t_wall_info *wall)
 {
 	double	*pos;
+	t_hlist	**textures_lst;
 
+	textures_lst = game->graphics->textures_lst;
 	if (wall->hit.tile == WALL)
 	{
-		wall->texture =
-			game->graphics->textures_lst[wall->hit.direction]->content;
+		wall->texture = textures_lst[wall->hit.direction]->content;
 		if (wall->hit.direction == NORTH || wall->hit.direction == WEST)
 		{
 			pos = (double *)malloc(sizeof(double) * 2);
