--- src/bonus/sprites/sprite_handler_bonus.c
@@ -1,12 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   sprite_handler_bonus.c                             :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: ana-cast <ana-cast@student.42malaga.com    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/05/26 19:37:55 by ana-cast          #+#    #+#             */
-/*   Updated: 2025/05/26 19:37:56 by ana-cast         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
